Move extracting field names to CPPUtil
